bot/requirements.txt
imagesize
numpy
opencv-python
portpicker
pyautogui
Pillow
PyQt6